connected feature flag page to home page through router
